requirements.txt

streamlit==1.29.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
pycocotools==2.0.7
nltk==3.8.1
tqdm==4.66.1
    